files and corresponding imports renamed from events. to event. and productions. to production.
