+++ backend/src/events/event.controller.ts
@@ -9,7 +9,7 @@ import {
   Put,
   UsePipes,
 } from "@nestjs/common";
-import { EventService } from "./events.service";
+import { EventService } from "./event.service";
 import { ZodValidationPipe } from "src/common/pipes/zod.validation.pipe";
 import { EventSchema, CreateEventSchema, UpdateEventSchema } from "@repo/common";
 import type { Event, CreateEvent, UpdateEvent } from "@repo/common";
