+++ backend/src/events/event.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
-import { EventService } from "./events.service";
-import { EventController } from "./events.controller";
+import { EventService } from "./event.service";
+import { EventController } from "./event.controller";
 import { EventDatabaseService } from "src/database/db.event.service";
 import { DbModule } from "../database/db.module";
