+++ backend/src/productions/production.module.ts
@@ -1,12 +1,12 @@
 import { Module } from "@nestjs/common";
-import { ProductionService } from "./productions.service";
-import { ProductionsController } from "./productions.controller";
+import { ProductionService } from "./production.service";
+import { ProductionController } from "./production.controller";
 import { ProductionDatabaseService } from "src/database/db.production.service";
 import { DbModule } from "../database/db.module";
 
 @Module({
   providers: [ProductionService],
-  controllers: [ProductionsController],
+  controllers: [ProductionController],
   imports: [DbModule],
 })
 export class ProductionModule {}
